CSS updates to improve layout. Added QAndAOptions component to separate buttons for user options at the bottom of the qAndA component.
