+++ client/src/components/qAndA/qAndAOptions.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const QAndAOptions = () => (
+  <div className="qAndA-options">
+    <button>More Answered Questions + </button>
+    <button>Add A Question</button>
+  </div>
+)
+
+export default QAndAOptions;
